docs: idea 3, added solve with T_GK
Added a description of the solution based on the Golub-Kahan matrix. Explained the method of inverse iterations and introduced a variation of the Thomas algorithm for counting called "non-monotonic Thomas" (the standard Thomas is monotonic in the sense that its forward and backward moves are sequential. The non-monotonic version is characterized by being more adaptive).